// src/db/generated/models/ContactRecord.ts
export type ContactRecordModel = runtime.Types.Result.DefaultSelection<Prisma.$ContactRecordPayload>
// ...
export type $ContactRecordPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "ContactRecord"
  objects: {
    caregiver: Prisma.$CaregiverPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    agencyId: string
    caregiverId: string
    line1: string | null
    line2: string | null
    city: string | null
    state: string | null
    zip: string | null
    mobilePhone: string | null
    alternatePhone: string | null
    email: string | null
    preferredLanguage: string | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["contactRecord"]>
  composites: {}
}