// src/db/generated/models/EmploymentEntry.ts
export type EmploymentEntryModel = runtime.Types.Result.DefaultSelection<Prisma.$EmploymentEntryPayload>
// ...
export type $EmploymentEntryPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "EmploymentEntry"
  objects: {
    caregiver: Prisma.$CaregiverPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    agencyId: string
    caregiverId: string
    employerName: string
    positionTitle: string | null
    supervisorName: string | null
    supervisorPhone: string | null
    line1: string | null
    line2: string | null
    city: string | null
    state: string | null
    zip: string | null
    startedOn: Date | null
    endedOn: Date | null
    isCurrent: boolean
    reasonForLeaving: string | null
    mayContact: boolean
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["employmentEntry"]>
  composites: {}
}